backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
diffusers
transformers
accelerate
lpips
Pillow
numpy
opencv-python-headless
pydantic
